Use the flow in predict module in Stan model
Attempted to mimic the flow in the predict module to build Stan
model. It seems however that this cannot be done because matrices in
Stan are real type while Poisson distribution is discrete so stan
cribs. Can't use int array either. Will try to avoid matrix
multiplication altogether in next try.