Multiple runs of disaggregation. Problems with viz logic
1. I know what the problem is with the disaggregation in the
log-linear model - the intercept is too big and the rate is pretty
much 0.
2. Messed up the viz logic. Basically we are using 3 data sets -
training, validation and projections. The problem is that 2 of them
have week of year (not dates) while the third one has dates, not week
of year. Not a big problem - just need to sort out a common x-axis for
the three.